feat: implementación de la gráfica lineal SensorPowerChart para el monitoreo de sensores y ajustes varios en la UI de la aplicación

+++ app/src/main/java/com/sensortv/app/model/SensorChartPoint.kt
@@ -4,6 +4,6 @@ package com.sensortv.app.model
  * Modelo de datos para representar un punto en el gráfico lineal de tiempo y potencia.
  */
 data class SensorChartPoint (
-    val timeStamp: Long,
+    val timeStamp: Float,
     val powerMw: Float
 )
